backend/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
